Fixed a "fix". Was clobbering the database every time, even if a separate user db had not been created (with GET).

+++ public_html/api/index.php
@@ -194,7 +194,7 @@
         }
         
         // Remove the file, if not passed as a parameter
-        if (file_exists($user_db) && !$_POST['db']) {
+        if ($_SERVER['REQUEST_METHOD'] === 'POST' && file_exists($user_db) && !$_POST['db']) {
             unlink($user_db);
         }
